refactor: use folder utils for helper classes and methods
