+++ python/src/admin/checks.py
@@ -27,7 +27,7 @@ from time import perf_counter, sleep
 import cv2
 from flask import Blueprint, redirect, render_template, url_for
 
-import upload
+import utils.upload as upload
 from admin.server_context import ctx
 from config import config
 from game_board.board import overlay_grid
