+++ python/src/scrabblewatch.py
@@ -16,7 +16,7 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <http://www.gnu.org/licenses/>.
 """
 
-from util import Static
+from utils.util import Static
 
 try:
     from hardware.oled import OLEDDisplay as DisplayImpl
